fix: build a rule selector from args, log it in list mode
